change name of GradientHelperData class, move it to separate file and docstring fixes
